+++ src/components/ui/ConfirmationModal.tsx
@@ -1,5 +1,5 @@
 import Button from '@/components/ui/Button';
-import Modal from '@/components/ui/Modal';
+import Modal from '@/components/ui/modal/Modal';
 import Text from '@/components/ui/Text';
 
 interface ConfirmationModalProps {
